Optimizations and Refactoring
Expression assembly frontend function reworked with the new str_to_op
function.
Assembly frontend file writes directly to a file instead of storing
assembly in a string to later write to a function. This should improve
performance as well as memory usage.
Added more fields to the AST struct which should improve performance and
memory usage as there is no copying of any heap allocated value.